Added winstates in rounds and games, as well as completed visuals
You can now win rounds, and rounds will go on as long as there are players in them.
When a game is over it will start over.
At the end of each round it will visually show who won and who is eliminated.
People who don't choose a throw will also be eliminated.
You are notified when you are eliminated and have to wait for the next round.
New rounds work as expected and people who are waiting are automatically added to the new round.